Move process permission sanity checks into mergePermissions
